moved plotcontrol to explorer instead of having a separate module
